+++ app.py
@@ -7,7 +7,6 @@ from flask_sqlalchemy import SQLAlchemy
 import hashlib
 
 from config import USER, PASSWORD, URL, PORT, DATABASE
-from Model import User
 from ext import redis
 
 app = Flask(__name__)
@@ -17,6 +16,22 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://{}:{}@{}:{}/{}'.format(USER, PA
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
 
 db = SQLAlchemy(app)
+
+
+class User(db.Model):
+
+    __tablename__ = 'user'
+
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(32), unique=True, index=True)
+    password = db.Column(db.String(128))
+    token = db.Column(db.String(128))
+
+    def __init__(self, name, pwd, token):
+        self.name = name
+        self.pwd = pwd
+        self.token = token
+        self.id = None
 
 
 @auth.verify_token
